Added Daniel Neilsen's arrayObject to Console
Daniel Neilsen submitted the arrayObject as a GarageGames resource for the Torque Game Engine in 2003. This add's a basic arrayObject which supports pushing, popping, moving, sorting, inserting, etc.

It may be worth either implementing this in the Torque 2D MIT project, or expanding the SimSet object to add more array support.