Hide the Cursor on Touch
This detects cursor behavior that could only be possible if the user has a touch screen and hides the cursor. Likewise, once the cursor is hidden, it detects behavior that could only be made by a mouse and shows the cursor again. The effect is that the cursor immediantly hides if the touch screen is used and shows again if the user grabs the mouse. Detection is necessary because in our experiments, windows 10 does not willing to tell us if the mouse or touch screen is used. This gives us a surefire way to recognize a touch screen even if the OS keeps it hidden. All of this can be controlled by a preference variable: $pref::Gui::hideCursorWhenTouchEventDetected.