Upgrades to GuiSpriteCtrl
The GuiSpriteControl has been upgraded to replace the GuiBitmap and GuiFadeInBitmap controls. It now has options to tile the image have a fixed aspect ratio. The image can be resized and offset and now has it's own image color instead of using the fill color. Finally it now has moveTo, GrowTo, and fadeTo functions that change offset, size, and color over time.