GuiGridCtrl Variable Height
This code allows child controls to have a variable height if the cell mode for that direction is set to variable. Previously only the width could be variable. Keep in mind that with the grid, "width" and "height" can be in either the vertical or horizontal direction. Yes, it's a little confusing.