GuiInspector with Profiles
This code updates the GuiInspector to have adjustable profiles instead of hard-coded profiles. It also allows various features of the inspector to be controlled in script when it is set up. The inspector itself has been changed to use a grid so that it can take better advantage of horizontal space. The label now defaults to directly above the control for each field. This can be changed as well in script. Not everything is perfect. More changes to come...