Hidden Field List for GuiInspector
The GuiInspector needed to be able to hide certain fields in certain contexts. For example, the Asset Manager doesn't show "Internal" assets which are used by the editors. As such, changing an asset to internal on accident would hide an asset forever. Other cases include when the editor would like to handle a field in a special way (like locked) or when changing a field is outside the scope to the editor.