Updates to GuiTextEditCtrl
This includes some serious updates to the GuiTextEditCtrl. The biggest change is that controls  now have an InputMode that can be used to require certain types of characters like plain text, a number, or a decimal value. There's also just a lot of clean up. There was an actual goto statement in there! It's gone now. I also removed a number of features that might have been cool 30 years ago, but seem a little out of place now. Script functions are now in their own file too.