SceneWindows with ScrollBars!
SceneWindows can now activate scrollbars to allow the user to easily move the camera around a scene. Although this might not be useful in every circumstance, it will certainly be useful in sprite based menus and editors. The AssetAdmin makes use of this by allowing the user to easily zoom in and scroll when looking at an asset by using the mouse wheel and scroll bars.