Update To GuiTextEditCtrl
This is a major update to the GuiTextEditCtrl. It allows the control to support multiline editing by turning on text wrap - removing the need for the GuiMLTextEditCtrl. It also changes the control to use std::string instead of the custom TextBuffer. Behavior has also been modernized. For example, double clicking a word highlights the word instead of all. Text boxes also now change on hover using the HL state. Two new values were added to the GuiControlProfile to handle the colors when text is selected. The only loss of functionality is that UTF16 strings are not supported, although this could be added at a later point.