Changes to UseInput
UseInput was a better-named version of the old "modal" field which worked backwards of the way it probably should have worked. I've moved it into the actual GuiControl and improved event handling by causing events to "bubble up" correctly. Generally speaking, UseInput should not be needed. Previously I had to pay close attention to it. Now it should mostly work as expected.