Canvas Find Mouse Control
This was a pretty major change which worked on the first try and doesn't seem to have any problems - so naturally I'm a little nervous that I'm missing something. With input events now bubbling up through the controls it was apparent that the old method of always having the old mouse control call leave before the new one calls enter is inadequate. If the mouse enters a child control it first "leaves" the parent despite the fact that clicking on the child might be  responded to by the parent. This change makes it so entering a child control does not leave a parent. As a result, you might call a dozen "enter" events to get to a child control without calling any leave events. As you move the the mouse out of the control stack, each control will then call its leave event correctly. I should also point out that enter and leave events do not bubble, because it really doesn't make sense to do so.