Project Manager Startup Screen
This adds a T2D startup screen that let's the user pick which "project" they want to load. Initially, the sandbox is the only project, but the user can make a new project using the startup screen. Projects are loosely defined as a folder off of the root that has an AppCore module in it. The startup screen uses the existing modules in the library folder to build the new project. There was a lot of work done around allowing a project to redefine named Gui Profiles since the editor needs these to be defined, but the projects will also need to define their own (to run without the editors).