FontSizeAdjust
With this change, the field FontSizeAdjust can be set on any GuiControl to change the font size for just that control. Rather than simply overriding the profile's FontSize, the FontSizeAdjust is a decimal value that is multiplied by the FontSize to produce a final, actual font size. So a FontSizeAdjust of 1 does nothing. If the FontSize is 12 and the FontSizeAdjust is 0.5, then the calulated font size is 6. This is an extremely useful feature that will eliminate the need to create profiles at every font size.