FontColor Override
With this code, a user can override the fontColor in a profile with their own color. This works for nearly all controls, except for a handful of editor specific controls and the GuiTextEditCtrl. To override a font color, simply set the FontColor on the control and then set the OverrideFontColor flag to true. That same flag can be used to turn off the override as well.