Fixed Relative Positioning Drift
I was continuing to see controls slowly drift out of position if relative positioning was turned on and you resized the window a few times. This was due to rounding that happens when the control's position changed. The fix effectively uses the first calculation in all future repositioning which prevents drift.