Project Manager Module Dependancies
This check-in adds support for syncing dependencies. Part of that was modifying the actual module syncing code to work how I wanted. So at this point, adding a module to a project will also add all of its dependencies. I also split the AppCore and Audio modules into a BlankGame module that the user can edit. Because they could get replaced, synced modules from the library should never be changed by the user.