Project Manager Launch on Startup
Added a checkbox for non-synced modules that states if the module should launch at startup. Behind the scenes this is done by setting the module's group to launch. The AppCore then loads the group at startup. Actually changing the group can only be done after the module is unregistered - so basically when the engine is shutting down.