Update consoleDictionary.h
This fixes a problem I ran into with modulus operator on negative numbers. Internally, a variable is stored as a signed float and an unsigned int. Unfortunately this is done using a conversion directly from a signed float to an unsigned int which gives different results depending on hardware. For whatever reason mine was saving a -1 (or 4294967295) as the int regardless of the actual value which was then accessed to do the modulus. This fixes the problem by converting the value first to a signed int and then to an unsigned int.