GuiEditor Foundation
This starts the process of moving the GuiEditor toy into the tools. It includes the GuiControl list and dragging controls onto the screen. More to come.

This also includes a few C++ bug fixes. Most notably, it fixes a bug that caused the GuiProfiles to have an incorrect reference count. Typically this error was on the side of having too many references (causing resources to be held onto for too long). Sometimes the ref count would error in the other direction and cause a fatal assert to fire. This has been fixed.