Update ChangeUsernameDlg.gui.taml
This minor change prevents the keyboard toy from crashing (although it still doesn't work). As we move forward, we will likely drop support for toys like this and move them into reusable modules that can be added to games so I don't want to spend too much effort repairing it.