Added Menus for the GuiEditors
This adds the menus needed for the GuiEditor. Most of the more important items don't work yet and the lesser items are inconsistent in how they work. I'll be addressing this in the future. The menu items also need to enable/disable based on the events of the editor. But at the very least, this code adds the menu. Baby steps.