Update to the GuiColorPickerCtrl
Repaired the various display modes. Changed the selector to be profile driven. Moved drawing functions to the dgl file like other drawing functions. Added checkers background to transparent part of the alpha slider. Added support for text and removed support for children controls. Moved script functions to a script file. Added "Ctrl" to the file names. Added a window to the Gui Editor that can test the different modes for the color picker (commented out for now).