GuiColorPopupCtrl
This adds a new color control to T2D! The GuiColorPopupCtrl is a fully functional color popup that needs very little configuration to use. Although you could easily build your own color popup using the GuiColorPickerCtrl, this one is ready to use and customizable enough to be useful in most projects.