Fixed Some Tab Issues
This fixes a number of issues with tabbing through the controls on the inspector. This adds a way to display when a check box and a drop down is the first responder. It also sets visibility of hidden direct children of an expand control to false when the expand control is collapsed and further makes it so non-visible controls cannot be tabbed to. It also fixes the focus code so that textboxes can lose focus when a click happens outside of the text box. When a textbox gains focus through tabbing, it now selects the contents of the box.