ImageAsset in Profile Bug
This latest version of the engine, finally keeps track of GuiProfile references correctly. As such, profiles actually drop their assets when they reach zero references (as originally intended). This revealed that the code to reinstate those assets was not actually working as hasn't been for a long time. The bug is fixed now.