GuiEditor and UseInput
The flag GuiControl.useInput allows a control to interact with an input device such as the keyboard or mouse. When off input is ignored by the control and its children. However, the flag should be ignored in the editor so that the control can still be clicked on and dragged around. This code fixes that.