GuiEditor and Named Objects
This causes the GuiEditor to avoid actually naming objects since doing so causes bugs if the object already exists in the engine. Instead it stores the name in a temporary editor name. As far as I can tell there are no side effects and hopefully it will stay that way.