SceneWindow Event Pass Through
This code allows sceneWindows to optionally let events that they don't consume pass through them to controls behind them. Like all events, sceneWindow events like onTouchDown can return true if they have consumed the event. Previously returning false would allow the parent of the sceneWindow to also respond to the event as it bubbles up the chain to the canvas. This bubbling will still happen, but afterwards the event, if still unconsumed, is passed to controls behind the sceneWindow. Practically, this allows two scene windows to be stacked so that the top window can be used for UI elements made from sprites and the bottom window can consist of a game world.