Complex Colors
This adds a features which I've named Complex Colors. Essentially it allows a separate blend color to be applied to each corner of a sprite. This allows much more complex blending scenarios that can mimic lighting. Complex colors can also be used with the individual sprites of a Composite Sprite.